requirements.txt
amadeus==9.0.0
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
colorama==0.4.6
flask==3.1.0
gunicorn==23.0.0
